Module minerva_rest::controller::auth

This submodule describes routes for authentication and session creation services.

Statics

AUTH_COOKIE

Cookie name for the authentication token, namely the session identification.

TENANT_COOKIE

Cookie name for the tenant name, saved within the browser.

Functions

get_endpoint

Retrieves the endpoint for the gRPC session service. Requires that the proper environment variables are defined.

login 🔒

Route for user login.

logout 🔒

Route for user logoff.

routes

Returns the list of routes for this module.