Function minerva_rest::controller::auth::login

async fn login(tenant: &str, body: Json<RecvSession>) -> Response

Route for user login.

This route requires that the tenant is informed on the login route. Furthermore, login needs login data for the creation of a session.

Upon a successful login attempt, the route will return the tenant and the session token data for the current user.

Request example

curl -X POST 'http://localhost:9000/minerva/login' \
     -H 'Content-Type: application/json' \
     -d '{"login": "admin", "password": "admin"}'