Module minerva_rest::controller::handlers

This submodule contains default status code catchers for errors within the API.

Structs

DefaultError 🔒

Default error structure implementing a Rocket responder for any status code.

Functions

bad_request 🔒

Error catcher for 400 - Bad Request.

catchers

Exports an array of catcher functions to be used when mounting the server.

default_catcher 🔒

Default error catcher for any kind of error that was not treated specifically.

gen_message 🔒

Generates a simple JSON string from a given message.

internal 🔒

Error catcher for 500 - Internal Server Error.

not_found 🔒

Error catcher for 404 - Not Found.

service_unavailable 🔒

Error catcher for 503 - Service Unavailable. Returned whenever a microservice cannot be reached to complete the request.

unauthorized 🔒

Error catcher for 401 - Unauthorized. Returned on problems with the authentication fairing.

unprocessable_entity 🔒

Error catcher for 412 - Unprocessable Entity. Returned when a JSON body to a request cannot be deserialized.