Function minerva_rest::controller::user::store

async fn store(
    tenant: String,
    session: SessionInfo,
    body: Json<RecvUser>
) -> Result<Json<User>, ErrorResponse>

Route for creating a new user.

To use this route, use a POST request, sending as body a JSON containing the expected data for creating a new user.

Upon success, returns the data for the created user as if it were requested through the show method.