Function minerva_rest::controller::user::store

async fn store(session: SessionInfo, body: Json<RecvUser>) -> Response

Route for creating a new user.

To use this route, use a POST request, sending as body a JSON containing the expected data for creating a new user.

Upon success, returns the data for the created user as if it were requested through the show method.

This route assumes that the authentication cookies are being passed as well.

Request example

curl -X POST 'http://localhost:9000/users' \
     -H 'Content-Type: application/json' \
     -b cookies.txt
     -d '{"login": "fulano", "name": "Fulano da Silva", "email": null, "password": "senha123"}'